Agregar desde git, list int

+++ ms-delivery/Controllers/WeatherForecastController.cs
@@ -28,7 +28,7 @@ namespace ms_delivery.Controllers
         {
             var rng = new Random();
 
-            var Listtest = new List<double>();
+            var Listtest = new List<int>();
 
 
             //cambio de emergencia
